+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the searcher sdks will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Rust: 0.13.0, Python 0.28.0, Javascript 0.29.0, api-types 0.12.0] - 2025-05-14
+
+- Failed bids for swap quotes now have a reason status that exposes why the transaction failed on-chain.
+
 ## [Rust: 0.12.0, Python 0.27.0, Javascript 0.28.0, api-types 0.11.0] - 2025-05-09
 
 - Add a profile ID field to the opportunity api type. [538](https://github.com/pyth-network/per/pull/538)